4_lightening_hydra_code_test_assignment/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
# --------- pytorch --------- #
timm==1.0.9
torch==2.4.0+cpu 
torchvision==0.19.0+cpu
lightning==2.4.0
torchmetrics==1.4.1
tensorboard==2.17.1
tensorboardX==2.6.2.2
scikit-learn==1.3.2

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
# hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
# pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
# pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- deployment --------- #
gradio==3.3.1

# ---------- pytest ------------- #
pytest==8.3.3 
pytest-cov==5.0.0


rootutils
loguru
